feat(tags): separate tags.$tag from tags.nav.$tags
