+++ app/routes/tags.$tag.tsx
@@ -1,14 +1,8 @@
 import {useState, useEffect, ReactNode} from 'react'
 import {LoaderFunction} from '@remix-run/cloudflare'
 import {reloadInBackgroundIfNeeded} from '~/server-utils/kv-cache'
-import {Tag as TagType, QuestionState, RelatedQuestions, loadTags} from '~/server-utils/stampy'
+import {Tag as TagType, QuestionState, RelatedQuestions, loadTag} from '~/server-utils/stampy'
 import Dialog from '~/components/dialog'
-import Footer from '~/components/Footer'
-import Header from '~/components/Header'
-import useToC from '~/hooks/useToC'
-import {useLoaderData} from '@remix-run/react'
-import {ListTable} from '~/components/Table/ListTable'
-import {CategoriesNav} from '~/components/CategoriesNav/Menu'
 
 type Props = {
   tags: string[]
@@ -23,15 +17,13 @@ export const loader = async ({request, params}: Parameters<LoaderFunction>[0]) =
   }
 
   try {
-    const tags = await loadTags(request)
-    return {...tags, tagFromUrl}
+    return await loadTag(request, tagFromUrl)
   } catch (error: unknown) {
     console.error(`error fetching tag "${tagFromUrl}":`, error)
     return {
       error: error?.toString(),
       timestamp: new Date().toISOString(),
-      data: new Array<TagType>(),
-      tagFromUrl,
+      data: [],
     }
   }
 }
@@ -48,11 +40,10 @@ export async function fetchTag(tagName: string): Promise<TagType | never[]> {
     const json: Awaited<ReturnType<typeof loader>> = await response.json()
     if ('error' in json) console.error(json.error)
     const {data, timestamp} = json
-    const currentTagData = data.filter((tagData) => tagData.name === tagName)[0]
 
     reloadInBackgroundIfNeeded(url, timestamp)
 
-    return currentTagData
+    return data
   })
 }
 
@@ -140,67 +131,5 @@ export function Tags({tags}: Props) {
     <div className="tags-container">
       <div>{tags && tags.map((name) => <Tag key={name} name={name} />)}</div>
     </div>
-  )
-}
-
-export default function App() {
-  const {tagFromUrl, data} = useLoaderData<ReturnType<typeof loader>>()
-  const [selectedTag, setSelectedTag] = useState<TagType | null>(null)
-  const [tagsFilter, setTagsFilter] = useState<string>('')
-  const {toc} = useToC()
-
-  const [sortBy] = useState<keyof typeof sortFuncs>('alphabetically')
-
-  useEffect(() => {
-    if (selectedTag === null) {
-      const dataForUrlTag = data.filter((tagData) => tagData.name === tagFromUrl)[0]
-      setSelectedTag(dataForUrlTag)
-    }
-  }, [selectedTag, data, tagFromUrl])
-  if (selectedTag === null) {
-    return null
-  }
-  return (
-    <>
-      <Header toc={toc} categories={data} />
-      <div className={'top-margin-large'} />
-      <main>
-        <div className={'group-elements'}>
-          <CategoriesNav
-            categories={
-              data
-                .filter((tag) => tag.questions.length > 0)
-                .filter((tag) => tag.name.toLowerCase().includes(tagsFilter.toLowerCase()))
-                .sort(sortFuncs[sortBy])
-
-              // {title: "AI Safety", id: 1},
-            }
-            active={selectedTag}
-            onClick={setSelectedTag}
-            onChange={setTagsFilter}
-          />
-
-          {selectedTag === null ? null : (
-            <div>
-              <h1 style={{marginTop: '0px'}}>{selectedTag.name}</h1>
-              {selectedTag.questions.length === 0 ? (
-                <div className={'no-questions'}>No questions found</div>
-              ) : (
-                <p>
-                  {selectedTag.questions.length} pages tagged {`"${selectedTag.name}"`}
-                </p>
-              )}
-              {selectedTag && <ListTable elements={selectedTag.questions} />}
-            </div>
-          )}
-        </div>
-      </main>
-
-      <div className={'top-margin-large-with-border'} />
-
-      <div className={'top-margin-large'} />
-
-      <Footer />
-    </>
   )
 }
